habdle octal conversion

+++ main.h
@@ -13,7 +13,7 @@ int _puts(char *str);
 int _printf(const char *format, ...);
 int choice(va_list r, char str);
 int printbin(long int num);
-int printoctal(int num);
+int printoctal(long int num);
 
 /**
  * struct choice - struct to nhold the choice of a person
